+++ src/dependency_risk_profiler/parsers/__init__.py
@@ -1,9 +1,23 @@
 """Package manifest parsers for different ecosystems."""
 
 from .base import BaseParser
+from .composer import ComposerLockParser
 from .golang import GoParser
+from .maven import MavenPomParser
 from .nodejs import NodeJSParser
+from .nuget import NuGetParser
 from .python import PythonParser
+from .ruby import GemfileLockParser
 from .toml import TomlParser
 
-__all__ = ["BaseParser", "GoParser", "NodeJSParser", "PythonParser", "TomlParser"]
+__all__ = [
+    "BaseParser",
+    "ComposerLockParser",
+    "GemfileLockParser",
+    "GoParser",
+    "MavenPomParser",
+    "NodeJSParser",
+    "NuGetParser",
+    "PythonParser",
+    "TomlParser",
+]